requirements.txt
# FastAPI и веб-сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Telegram бот
aiogram==3.3.0
aiohttp==3.9.1

# Парсинг документов
python-docx==1.1.0

# Планировщик задач
apscheduler==3.10.4

# Утилиты
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Генерация кодов и безопасность
secrets
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Работа с датами
pytz==2023.3

# Логирование
loguru==0.7.2

# Excel экспорт (для отчетов)
openpyxl==3.1.2
pandas==2.1.4
